+++ src/AppiumLibrary/keywords/_applicationmanagement.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import os
+import robot
 from appium import webdriver
 from AppiumLibrary.utils import ApplicationCache
 from keywordgroup import KeywordGroup
@@ -9,7 +10,6 @@ ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 
 
 class _ApplicationManagementKeywords(KeywordGroup):
-
     def __init__(self):
         self._cache = ApplicationCache()
         self._timeout_in_secs = float(5)
@@ -35,9 +35,9 @@ class _ApplicationManagementKeywords(KeywordGroup):
         self._debug('Closing all applications')
         self._cache.close_all()
 
-    def open_application(self, remote_url, alias=None,  **kwargs):
+    def open_application(self, remote_url, alias=None, **kwargs):
         """Opens a new application to given Appium server.
-        Capabilities of appium server, Android and iOS, 
+        Capabilities of appium server, Android and iOS,
         Please check http://appium.io/slate/en/master/?python#appium-server-capabilities
         | *Option*            | *Man.* | *Description*     |
         | remote_url          | Yes    | Appium server url |
@@ -49,8 +49,9 @@ class _ApplicationManagementKeywords(KeywordGroup):
         """
         desired_caps = kwargs
         application = webdriver.Remote(str(remote_url), desired_caps)
+
         self._debug('Opened application with session id %s' % application.session_id)
-        
+
         return self._cache.register(application, alias)
 
     def switch_application(self, index_or_alias):
@@ -79,9 +80,50 @@ class _ApplicationManagementKeywords(KeywordGroup):
             self._cache.switch(index_or_alias)
         return old_index
 
+    def reset_application(self):
+        """ Reset application """
+        driver = self._current_application()
+        driver.reset()
+
+    def remove_application(self, application_id):
+        """ Removes the application that is identified with an application id
+
+        Example:
+        | Remove Application |  com.netease.qa.orangedemo |
+
+        """
+        driver = self._current_application()
+        driver.remove_app(application_id)
+
+    def get_appium_timeout(self):
+        """Gets the timeout in seconds that is used by various keywords.
+
+        See `Set Appium Timeout` for an explanation."""
+        return robot.utils.secs_to_timestr(self._timeout_in_secs)
+
+    def set_appium_timeout(self, seconds):
+        """Sets the timeout in seconds used by various keywords.
+
+        There are several `Wait ...` keywords that take timeout as an
+        argument. All of these timeout arguments are optional. The timeout
+        used by all of them can be set globally using this keyword.
+
+        The previous timeout value is returned by this keyword and can
+        be used to set the old value back later. The default timeout
+        is 5 seconds, but it can be altered in `importing`.
+
+        Example:
+        | ${orig timeout} = | Set Appium Timeout | 15 seconds |
+        | Open page that loads slowly |
+        | Set Appium Timeout | ${orig timeout} |
+        """
+        old_timeout = self.get_appium_timeout()
+        self._timeout_in_secs = robot.utils.timestr_to_secs(seconds)
+        return old_timeout
+
     def get_source(self):
         """Returns the entire source of the current page."""
-        return self._current_application().page_source    
+        return self._current_application().page_source
 
     def log_source(self, loglevel='INFO'):
         """Logs and returns the entire html source of the current page or frame.
@@ -116,6 +158,18 @@ class _ApplicationManagementKeywords(KeywordGroup):
         """
         self._current_application().shake()
 
+    def portrait(self):
+        """
+        Set the device orientation to PORTRAIT
+        """
+        self._rotate('PORTRAIT')
+
+    def landscape(self):
+        """
+        Set the device orientation to LANDSCAPE
+        """
+        self._rotate('LANDSCAPE')
+
     def get_current_context(self):
         """Get current context."""
         return self._current_application().current_context
@@ -128,11 +182,11 @@ class _ApplicationManagementKeywords(KeywordGroup):
     def switch_to_context(self, context_name):
         """Switch to a new context"""
         self._current_application().switch_to.context(context_name)
-        
+
     def go_to_url(self, url):
         """
-        Opens URL in default web browser. 
-        
+        Opens URL in default web browser.
+
         Example:
         | Open Application  | http://localhost:4755/wd/hub | platformName=iOS | platformVersion=7.0 | deviceName='iPhone Simulator' | browserName=Safari |
         | Go To URL         | http://m.webapp.com          |
@@ -162,3 +216,7 @@ class _ApplicationManagementKeywords(KeywordGroup):
 
     def _is_andriod(self):
         return self._is_platform('android')
+
+    def _rotate(self, orientation):
+        driver = self._current_application()
+        driver.orientation = orientation