+++ src/AppiumLibrary/keywords/_element.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
 
-from appium.webdriver.common.touch_action import TouchAction
-from AppiumLibrary import utils
 from AppiumLibrary.locators import ElementFinder
 from keywordgroup import KeywordGroup
+from robot.libraries.BuiltIn import BuiltIn
+
 
 class _ElementKeywords(KeywordGroup):
-
     def __init__(self):
         self._element_finder = ElementFinder()
 
@@ -26,12 +25,12 @@ class _ElementKeywords(KeywordGroup):
         `introduction` for details about locating elements.
         """
         self._info("Clicking element '%s'." % locator)
-        self._element_find(locator, True, True).click()        
+        self._element_find(locator, True, True).click()
 
     def click_button(self, index_or_name):
         """ Click button """
-        _platform_class_dict = {'ios':'UIAButton',
-                                   'android': 'android.widget.Button'}
+        _platform_class_dict = {'ios': 'UIAButton',
+                                'android': 'android.widget.Button'}
         if self._is_support_platform(_platform_class_dict):
             class_name = self._get_class(_platform_class_dict)
             self._click_element_by_class_name(class_name, index_or_name)
@@ -62,19 +61,12 @@ class _ElementKeywords(KeywordGroup):
         self._info("Setting text '%s' into text field '%s'" % (text, locator))
         self._element_input_value_by_locator(locator, text)
 
-    def reset_application(self):
-        """ Reset application """
-        driver = self._current_application()
-        driver.reset()
-
-    def hide_keyboard(self):
-        """
-        Hides the software keyboard on the device, using the specified key to
-        press. If no key name is given, the keyboard is closed by moving focus
-        from the text field. iOS only.
+    def hide_keyboard(self, key_name=None):
+        """Hides the software keyboard on the device. (optional) In iOS, use `key_name` to press
+        a particular key, ex. `Done`. In Android, no parameters are used.
         """
         driver = self._current_application()
-        driver.hide_keyboard()
+        driver.hide_keyboard(key_name)
 
     def page_should_contain_text(self, text, loglevel='INFO'):
         """Verifies that current page contains `text`.
@@ -83,7 +75,7 @@ class _ElementKeywords(KeywordGroup):
         using the log level specified with the optional `loglevel` argument.
         Giving `NONE` as level disables logging.
         """
-        if not text in self.log_source(loglevel):
+        if text not in self.log_source(loglevel):
             self.log_source(loglevel)
             raise AssertionError("Page should have contained text '%s' "
                                  "but did not" % text)
@@ -108,11 +100,11 @@ class _ElementKeywords(KeywordGroup):
         If this keyword fails, it automatically logs the page source
         using the log level specified with the optional `loglevel` argument.
         Givin
-        """        
+        """
         if not self._is_element_present(locator):
             self.log_source(loglevel)
             raise AssertionError("Page should have contained element '%s' "
-                                 "but did not" % locator)            
+                                 "but did not" % locator)
         self._info("Current page contains element '%s'." % locator)
 
     def page_should_not_contain_element(self, locator, loglevel='INFO'):
@@ -121,14 +113,14 @@ class _ElementKeywords(KeywordGroup):
         If this keyword fails, it automatically logs the page source
         using the log level specified with the optional `loglevel` argument.
         Givin
-        """   
+        """
         if self._is_element_present(locator):
             self.log_source(loglevel)
             raise AssertionError("Page should not have contained element '%s' "
-                                 "but did not" % locator)            
+                                 "but did not" % locator)
         self._info("Current page not contains element '%s'." % locator)
 
-    def element_should_be_disabled(self, locator):
+    def element_should_be_disabled(self, locator, loglevel='INFO'):
         """Verifies that element identified with locator is disabled.
 
         Key attributes for arbitrary elements are `id` and `name`. See
@@ -137,10 +129,10 @@ class _ElementKeywords(KeywordGroup):
         if self._element_find(locator, True, True).is_enabled():
             self.log_source(loglevel)
             raise AssertionError("Element '%s' should be disabled "
-                                 "but did not" % locator)      
+                                 "but did not" % locator)
         self._info("Element '%s' is disabled ." % locator)
 
-    def element_should_be_enabled(self, locator):
+    def element_should_be_enabled(self, locator, loglevel='INFO'):
         """Verifies that element identified with locator is enabled.
 
         Key attributes for arbitrary elements are `id` and `name`. See
@@ -149,24 +141,41 @@ class _ElementKeywords(KeywordGroup):
         if not self._element_find(locator, True, True).is_enabled():
             self.log_source(loglevel)
             raise AssertionError("Element '%s' should be enabled "
-                                 "but did not" % locator)      
+                                 "but did not" % locator)
         self._info("Element '%s' is enabled ." % locator)
-
 
     def element_name_should_be(self, locator, expected):
         element = self._element_find(locator, True, True)
         if expected != element.get_attribute('name'):
             raise AssertionError("Element '%s' name should be '%s' "
-                                 "but it is '%s'." % (locator, expected, element.get_attribute('name'))) 
+                                 "but it is '%s'." % (locator, expected, element.get_attribute('name')))
         self._info("Element '%s' name is '%s' " % (locator, expected))
 
-    def element_attribute_should_be(self, locator, attr_name, expected):
+    def element_value_should_be(self, locator, expected):
+        element = self._element_find(locator, True, True)
+        if expected != element.get_attribute('value'):
+            raise AssertionError("Element '%s' value should be '%s' "
+                                 "but it is '%s'." % (locator, expected, element.get_attribute('value')))
+        self._info("Element '%s' value is '%s' " % (locator, expected))
+
+    def element_attribute_should_match(self, locator, attr_name, match_pattern, regexp = False):
         elements = self._element_find(locator, False, True)
         if len(elements) > 1:
             self._info("CAUTION: '%s' matched %s elements - using the first element only" % (locator, len(elements)))
-        if expected != elements[0].get_attribute(attr_name):
-            raise AssertionError("Element '%s' attribute '%s' should have been '%s' "
-                                 "but it was '%s'." % (locator, attr_name, expected, element.get_attribute('name')))
+        attr_value = elements[0].get_attribute(attr_name)
+        if regexp:
+            BuiltIn.should_match_regexp(self,attr_value,match_pattern,
+                                        msg="Element '%s' attribute '%s' should have been '%s' "
+                                        "but it was '%s'." % (locator, attr_name, expected, attr_value),
+                                        values=False)
+        else:
+            BuiltIn.should_match(self,attr_value,match_pattern,
+                                        msg="Element '%s' attribute '%s' should have been '%s' "
+                                        "but it was '%s'." % (locator, attr_name, expected, attr_value),
+                                        values=False)
+        #if expected != elements[0].get_attribute(attr_name):
+        #    raise AssertionError("Element '%s' attribute '%s' should have been '%s' "
+        #                         "but it was '%s'." % (locator, attr_name, expected, element.get_attribute(attr_name)))
         self._info("Element '%s' attribute '%s' is '%s' " % (locator, attr_name, expected))
 
     def get_element_attribute(self, locator, attr_name):
@@ -178,9 +187,31 @@ class _ElementKeywords(KeywordGroup):
     def get_elements(self, locator, first_element_only=False, fail_on_error=True):
         return self._element_find(locator, first_element_only, fail_on_error)
 
+    def get_element_attribute(self, locator, attribute):
+        """Get element attribute using given attribute: name, value,...
+
+        Example:
+
+        |Get Element Attribute| locator | name |
+        or
+        |Get Element Attribute| locator | value |
+        """
+        elements = self._element_find(locator, False, True)
+        ele_len = len(elements)
+        if ele_len == 0:
+            raise AssertionError("Element '%s' could not be found" % locator)
+        elif ele_len > 1:
+            self._info("CAUTION: '%s' matched %s elements - using the first element only" % (locator, len(elements)))
+
+        try:
+            attr_val = element.get_attribute(attribute)
+            self._info("Element '%s' attribute '%s' value '%s' " % (locator, attribute, attr_val))
+            return attr_val
+        except:
+            raise AssertionError("Attribute '%s' is not valid for element '%s'" % (attribute, locator))
 
     # Private
-    
+
     def _is_index(self, index_or_name):
         if index_or_name.startswith('index='):
             return True
@@ -193,7 +224,7 @@ class _ElementKeywords(KeywordGroup):
             element = driver.find_element_by_name(name)
         except Exception, e:
             raise Exception, e
-    
+
         try:
             element.click()
         except Exception, e:
@@ -206,12 +237,12 @@ class _ElementKeywords(KeywordGroup):
 
     def _find_element_by_class_name(self, class_name, index_or_name):
         elements = self._find_elements_by_class_name(class_name)
-    
+
         if self._is_index(index_or_name):
             try:
                 index = int(index_or_name.split('=')[-1])
                 element = elements[index]
-            except IndexError, TypeError:
+            except (IndexError, TypeError):
                 raise Exception, 'Cannot find the element with index "%s"' % index_or_name
         else:
             found = False
@@ -226,7 +257,7 @@ class _ElementKeywords(KeywordGroup):
         return element
 
     def _get_class(self, platform_class_dict):
-        return platform_class_dict.get(self._get_platform())        
+        return platform_class_dict.get(self._get_platform())
 
     def _is_support_platform(self, platform_class_dict):
         return platform_class_dict.has_key(self._get_platform())