+++ src/AppiumLibrary/keywords/_logging.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 
 import os
-import sys
-from robot.variables import GLOBAL_VARIABLES
+from robot.libraries.BuiltIn import BuiltIn
 from robot.api import logger
 from keywordgroup import KeywordGroup
+
 
 class _LoggingKeywords(KeywordGroup):
 
@@ -14,10 +14,11 @@ class _LoggingKeywords(KeywordGroup):
         logger.debug(message)
 
     def _get_log_dir(self):
-        logfile = GLOBAL_VARIABLES['${LOG FILE}']
+        variables = BuiltIn().get_variables()
+        logfile = variables['${LOG FILE}']
         if logfile != 'NONE':
             return os.path.dirname(logfile)
-        return GLOBAL_VARIABLES['${OUTPUTDIR}']
+        return variables['${OUTPUTDIR}']
 
     def _html(self, message):
         logger.info(message, True, False)
@@ -27,13 +28,17 @@ class _LoggingKeywords(KeywordGroup):
 
     def _log(self, message, level='INFO'):
         level = level.upper()
-        if (level == 'INFO'): self._info(message)
-        elif (level == 'DEBUG'): self._debug(message)
-        elif (level == 'WARN'): self._warn(message)
-        elif (level == 'HTML'): self._html(message)
+        if (level == 'INFO'):
+            self._info(message)
+        elif (level == 'DEBUG'):
+            self._debug(message)
+        elif (level == 'WARN'):
+            self._warn(message)
+        elif (level == 'HTML'):
+            self._html(message)
 
     def _log_list(self, items, what='item'):
-        msg = ['Altogether %d %s%s.' % (len(items), what, ['s',''][len(items)==1])]
+        msg = ['Altogether %d %s%s.' % (len(items), what, ['s', ''][len(items) == 1])]
         for index, item in enumerate(items):
             msg.append('%d: %s' % (index+1, item))
         self._info('\n'.join(msg))