+++ src/AppiumLibrary/keywords/_screenshot.py
@@ -4,6 +4,7 @@ import os
 import robot
 from keywordgroup import KeywordGroup
 
+
 class _ScreenshotKeywords(KeywordGroup):
 
     def __init__(self):
@@ -28,9 +29,9 @@ class _ScreenshotKeywords(KeywordGroup):
         path, link = self._get_screenshot_paths(filename)
 
         if hasattr(self._current_application(), 'get_screenshot_as_file'):
-          self._current_application().get_screenshot_as_file(path)
+            self._current_application().get_screenshot_as_file(path)
         else:
-          self._current_application().save_screenshot(path)
+            self._current_application().save_screenshot(path)
 
         # Image is shown on its own row and thus prev row is closed on purpose
         self._html('</td></tr><tr><td colspan="3"><a href="%s">'